feat: use the new ad event parameter (#411)
#### What problem is this solving?

<!--- What is the motivation and context for this change? -->
We're changing the encoded event parameters from `adId` to
`eventParameters`, since it's not the ad ID. The data attribute is still
using the data-van-aid since that's what the Activity Flow use to track
events (to be changed in future).

This is backwards compatible as it keeps using the `adId` as fallback to
the `eventParameters` parameter.

It also removes many data properties that are not being used anymore,
since the `data-van-aid` is the only one considered by Activity Flow
currentyl

#### How to test it?

<!--- Don't forget to add a link to a Workspace where this branch is
linked -->
In this workspace, https://leal--biggy.myvtex.com/ search for the term
"camisa".



#### Screenshots or example usage:

<img width="1624" height="1060" alt="image"
src="https://github.com/user-attachments/assets/3593a558-1ac6-4cb8-859c-93fb6fea0277"
/>

<!--- Add some images or gifs to showcase changes in behaviour or
layout. Example: before and after images -->


#### Related to / Depends on

https://github.com/vtex-apps/store/pull/605

https://github.com/vtex-apps/search-result/pull/711

---------

Co-authored-by: Matheus Leal <matheus.leal@vtex.com>
Co-authored-by: Vinícius Seixas <vinicius.seixas@vtex.com>